fixed / refactored tests
