+++ tests/test_builtins.py
@@ -1,11 +1,8 @@
-# -*- coding: utf-8 -*-
 """Tests the xonsh builtins."""
-from __future__ import unicode_literals, print_function
 import os
 import re
 import types
 from ast import AST, Module, Interactive, Expression
-from subprocess import Popen
 
 import pytest
 
@@ -23,10 +20,8 @@ from xonsh.built_ins import (
     in_macro_call,
     call_macro,
     enter_macro,
-    cmds_to_specs,
 )
 from xonsh.environ import Env
-from xonsh.proc import PopenThread, ProcProxy, ProcProxyThread
 
 from tools import skip_if_on_windows
 
@@ -388,28 +383,3 @@ def test_enter_macro():
     assert obj.macro_block == "wakka"
     assert obj.macro_globals
     assert obj.macro_locals
-
-
-@skip_if_on_windows
-def test_cmds_to_specs_thread_subproc(xonsh_builtins):
-    env = xonsh_builtins.__xonsh__.env
-    cmds = [["pwd"]]
-    # First check that threadable subprocs become threadable
-    env["THREAD_SUBPROCS"] = True
-    specs = cmds_to_specs(cmds, captured="hiddenobject")
-    assert specs[0].cls is PopenThread
-    # turn off threading and check we use Popen
-    env["THREAD_SUBPROCS"] = False
-    specs = cmds_to_specs(cmds, captured="hiddenobject")
-    assert specs[0].cls is Popen
-
-    # now check the threadbility of callable aliases
-    cmds = [[lambda: "Keras Selyrian"]]
-    # check that threadable alias become threadable
-    env["THREAD_SUBPROCS"] = True
-    specs = cmds_to_specs(cmds, captured="hiddenobject")
-    assert specs[0].cls is ProcProxyThread
-    # turn off threading and check we use ProcProxy
-    env["THREAD_SUBPROCS"] = False
-    specs = cmds_to_specs(cmds, captured="hiddenobject")
-    assert specs[0].cls is ProcProxy
