+++ xonsh/procs/readers.py
@@ -4,6 +4,7 @@ import io
 import sys
 import time
 import queue
+import ctypes
 import builtins
 import threading
