+++ xonsh/xoreutils/cat.py
@@ -4,7 +4,7 @@ import sys
 import time
 import builtins
 
-import xonsh.proc as xproc
+import xonsh.procs.pipelines as xpp
 from xonsh.xoreutils.util import arg_handler
 
 
@@ -57,7 +57,7 @@ def _cat_single_file(opts, fname, stdin, out, err, line_count=1):
         if file_size == 0:
             file_size = None
         fobj = open(fname, "rb")
-        f = xproc.NonBlockingFDReader(fobj.fileno(), timeout=0.1)
+        f = xpp.NonBlockingFDReader(fobj.fileno(), timeout=0.1)
     sep = os.linesep.encode(enc, enc_errors)
     last_was_blank = False
     while file_size is None or read_size < file_size:
